Squashed commits:
[0601e30] Improve Matter Furnace automations
Previously sides couldn't be used at all for input, and the only input was into the fuel slot from the top/bottom.
This meant we couldn't automatically input stuff to smelt at all!
This has been changed as follows
- Extraction: All sides, only from output slots
- Insertion: Top = Input; Sides = Fuels; Bottom = Disallowed.
This matches the vanilla furnace exactly, except that extraction is allowed from all sides.

[4b63b45] Small improvement to collector automation interaction
Items in the output slot are now held there if they match the lock item.
This allows automation a chance to extract before it gets dumped back into the input