One True Code style has now been applied.
Long ago, in a land far, far away, the lord of same-line braces was born. He looked upon his coding minions, and smiled in favor.
Then came a dark time. The anti-lord of next-line braces rose in the west. It looked upon its minions, and crushed them underfoot, forcing them to adopt its coding style.
Now, we revolt. We throw off the shackles of next-line-dom in a commit that TOUCHES EVERYTHING. None shall escape liberation and restoration to the pristine state of the Before-Time.
May the brace be forever on the same line for you, fellow developer.

Signed this 1st of April,
williewillus

(if this actually gets pulled I will laugh my ass off)